pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit-core>=3.2,<4"]

[project]
name = "aa-ephemeral"
dynamic = ["description", "version"]
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
dependencies = [
    "allianceauth>=3,<6",
    "allianceauth-discordbot>=5,<6",
    "py-cord>2,<3",
]

[tool.flit.module]
name = "aa_ephemeral"
